+++ config/routes.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal:true
 
 Rails.application.routes.draw do
-  get 'explore_collections', to: 'explore_collections#index'
+  get 'explore_collections', to: 'oregon_digital/explore_collections#index'
 
   namespace :admin do
     resources :collection_types, except: :show, controller: 'oregon_digital/collection_types'
