+++ spec/controllers/oregon_digital/explore_collections_controller_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal:true
+
+RSpec.describe OregonDigital::ExploreCollectionsController do
+  describe '#index' do
+    subject { get :index, format: 'text/html' }
+
+    it { s_expected.to render_template('index') }
+  end
+end
